!!breaking change: invert the dependency between the generated API spec and client
The AppStoreConnect module no longer forwards imports from the AppStoreAPI target. The client remains in `AppStoreConnect`, and going forward you will need to also import `AppStoreAPI` if you wish to use the App Store Connect API requests. This is to enable builds of the Enterprise Program API that do not depend on ASC specific functionality, and vis-a-versa.